feat: implement dynamic door name placement with collision avoidance
This change introduces a force-directed collision resolution system for
door labels rendered on the map. Labels now try to avoid overlapping
each other by shifting their positions while maintaining a spring force
towards their original anchor points.

Key changes:
- Refactored `FontMetrics` and `GLText` into dedicated headers to
  allow text measurement in background threads.
- Implemented an O(N) force-directed solver using a spatial hash grid
  in `MapCanvasRoomDrawer.cpp`.
- Integrated collision resolution into the asynchronous mesh generation
  pipeline, baking optimized positions into the final font mesh.
- Optimized label batching to eliminate redundant string copies.
- Fixed header dependencies and potential circular includes.

The solver uses a reference zoom of 0.4 to calculate bounding boxes,
ensuring legibility even when zoomed out. Perfectly overlapping labels
are handled via a tiny random jitter to initiate separation.
